requirements.txt
python-socketio[client]
langchain
openai
pinecone-client
python-dotenv
asyncio
PyPDF2